vmtests: add s390x config
Unlike the existing x86_64 config, this is a sorted defconfig, which is
more convenient to diff and to work with. make olddefconfig processes it
just fine. For now CONFIG_DEBUG_INFO_BTF is disabled, because it doesn't
support cross-compilation for a different endianness.

Also, select the current config based on LIBBPF_ARCH value and thus
rename the existing config to latest-x86_64.config.

Signed-off-by: Ilya Leoshkevich <iii@linux.ibm.com>